add tlut calculation
