+++ lib/AuroraLip/Palette/Formats/PLT0.cs
@@ -1,4 +1,5 @@
 ﻿using AuroraLip.Common;
+using AuroraLip.Texture;
 using System;
 using System.Collections.Generic;
 using System.IO;
